Make arlexecute as needed for tests
