+++ tests/workflows/test_skymodel_arlexecute.py
@@ -13,7 +13,7 @@ from data_models.memory_data_models import Image
 from data_models.memory_data_models import Skycomponent
 from data_models.polarisation import PolarisationFrame
 from workflows.arlexecute.skymodel.skymodel_arlexecute import predict_skymodel_list_arlexecute_workflow
-from wrappers.arlexecute.execution_support.arlexecute import arlexecute
+from wrappers.arlexecute.execution_support.arlexecutebase import ARLExecuteBase
 from wrappers.arlexecute.execution_support.dask_init import get_dask_Client
 from wrappers.arlexecute.simulation.testing_support import ingest_unittest_visibility, \
     create_low_test_skymodel_from_gleam
@@ -30,13 +30,18 @@ class TestSkyModel(unittest.TestCase):
     def setUp(self):
         
         client = get_dask_Client(memory_limit=4 * 1024 * 1024 * 1024, n_workers=4, dashboard_address=None)
+        global arlexecute
+        arlexecute = ARLExecuteBase(use_dask=True)
         arlexecute.set_client(client, verbose=True)
         
         from data_models.parameters import arl_path
         self.dir = arl_path('test_results')
         
     def tearDown(self):
+        global arlexecute
         arlexecute.close()
+        del arlexecute
+
     
     def actualSetUp(self, freqwin=1, block=False, dopol=False, zerow=False):
